move mysql_config test from dm/worker to dm/config
