+++ dm/worker/config.go
@@ -33,10 +33,6 @@ import (
 // and assign it to SampleConfigFile while we build dm-worker
 var SampleConfigFile string
 
-var (
-	getAllServerIDFunc = utils.GetAllServerID
-)
-
 // NewConfig creates a new base config for worker.
 func NewConfig() *Config {
 	cfg := &Config{}
